# app/handlers/auth.go
/*
 * Обработка запросов авторизации
 */
package handlers

import (
	"fmt"

	entity "museum/app/entity/auth"
	"museum/app/handlers/helpers"
	models "museum/app/models/user"

	"github.com/go-playground/validator/v10"
	"github.com/gofiber/fiber/v2"
)

type authRoutes struct{}

// Подключение роутов обработки авторизации
func NewAuthRoutes() *authRoutes {
	return &authRoutes{}
}

// SignUp godoc
// @Summary      Регистрация
// @Description  Ввод почты, пароля и подтверждение пароля
// @Tags         Авторизация
// @Accept       json
// @Produce      json
// @Param        id   path      int  true  "Account ID"
// @Success      200
// @Router       /v1/auth/sign_up [post]
func (r *authRoutes) SignUp(ctx *fiber.Ctx) error {
	var req entity.SignUpEntity
	if err := ctx.BodyParser(&req); err != nil {
		return fmt.Errorf("body parser: %w", err)
	}

	err := validator.New().Struct(req)
	if err != nil {
		errors := err.(validator.ValidationErrors)
		fmt.Println(errors.Error())
		return ctx.Status(fiber.StatusUnprocessableEntity).JSON(err.Error())
	}

	token, err := helpers.GenerateToken("1", "email")
	if err != nil {
		return ctx.SendStatus(fiber.StatusInternalServerError)
	}
	return ctx.Status(fiber.StatusCreated).JSON(fiber.Map{"token": token})
}

// SignIn godoc
// @Summary      Вход
// @Description  Проверка почты и пароля, генерация токена
// @Tags         Авторизация
// @Accept       json
// @Produce      json
// @Param        q    query     string  false  "name search by q"  Format(email)
// @Success      200
// @Router       /v1/auth/sign_in [post]
func (r *authRoutes) SignIn(ctx *fiber.Ctx) error {
	// user := models.NewUser()
	token, err := helpers.GenerateToken("1", "email")
	fmt.Printf("%s", err)
	if err != nil {
		return ctx.SendStatus(fiber.StatusInternalServerError)
	}
	return ctx.Status(fiber.StatusAccepted).JSON(fiber.Map{"token": token})
}

// GetMe godoc
// @Summary      Получение текущего пользователя
// @Description  Поиск пользователя по токену
// @Tags         Авторизация
// @Accept       json
// @Produce      json
// @Param        q    query     string  false  "name search by q"  Format(email)
// @Success      200
// @Router       /v1/auth/me [get]
func (r *authRoutes) GetMe(ctx *fiber.Ctx) error {
	user := ctx.Locals("currentUser").(*models.User)
	return ctx.Status(fiber.StatusGone).JSON(fiber.Map{"email": user.Email})
}